+++ server/v2/cometbft/commands.go
@@ -19,17 +19,16 @@ import (
 	"sigs.k8s.io/yaml"
 
 	"cosmossdk.io/server/v2/cometbft/client/rpc"
-	"cosmossdk.io/server/v2/cometbft/flags"
 	auth "cosmossdk.io/x/auth/client/cli"
-	"github.com/cosmos/cosmos-sdk/client"
 
+	"github.com/cosmos/cosmos-sdk/client"
 	cryptocodec "github.com/cosmos/cosmos-sdk/crypto/codec"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/query"
 	"github.com/cosmos/cosmos-sdk/version"
 )
 
-func (s *CometBFTServer[T]) rpcClient(cmd *cobra.Command) (rpc.CometRPC, error) {
+func (s *CometBFTServer[AppT, T]) rpcClient(cmd *cobra.Command) (rpc.CometRPC, error) {
 	if s.config.Standalone {
 		client, err := rpchttp.New(client.GetConfigFromCmd(cmd).RPC.ListenAddress)
 		if err != nil {
@@ -38,8 +37,8 @@ func (s *CometBFTServer[T]) rpcClient(cmd *cobra.Command) (rpc.CometRPC, error) 
 		return client, nil
 	}
 
-	if s.Node == nil || cmd.Flags().Changed(flags.FlagNode) {
-		rpcURI, err := cmd.Flags().GetString(flags.FlagNode)
+	if s.Node == nil || cmd.Flags().Changed(FlagNode) {
+		rpcURI, err := cmd.Flags().GetString(FlagNode)
 		if err != nil {
 			return nil, err
 		}
@@ -52,7 +51,7 @@ func (s *CometBFTServer[T]) rpcClient(cmd *cobra.Command) (rpc.CometRPC, error) 
 }
 
 // StatusCommand returns the command to return the status of the network.
-func (s *CometBFTServer[T]) StatusCommand() *cobra.Command {
+func (s *CometBFTServer[AppT, T]) StatusCommand() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "status",
 		Short: "Query remote node for status",
@@ -76,14 +75,14 @@ func (s *CometBFTServer[T]) StatusCommand() *cobra.Command {
 		},
 	}
 
-	cmd.Flags().StringP(flags.FlagNode, "n", "tcp://localhost:26657", "Node to connect to")
-	cmd.Flags().StringP(flags.FlagOutput, "o", "json", "Output format (text|json)")
+	cmd.Flags().StringP(FlagNode, "n", "tcp://localhost:26657", "Node to connect to")
+	cmd.Flags().StringP(FlagOutput, "o", "json", "Output format (text|json)")
 
 	return cmd
 }
 
 // ShowNodeIDCmd - ported from CometBFT, dump node ID to stdout
-func (s *CometBFTServer[T]) ShowNodeIDCmd() *cobra.Command {
+func (s *CometBFTServer[AppT, T]) ShowNodeIDCmd() *cobra.Command {
 	return &cobra.Command{
 		Use:   "show-node-id",
 		Short: "Show this node's ID",
@@ -101,7 +100,7 @@ func (s *CometBFTServer[T]) ShowNodeIDCmd() *cobra.Command {
 }
 
 // ShowValidatorCmd - ported from CometBFT, show this node's validator info
-func (s *CometBFTServer[T]) ShowValidatorCmd() *cobra.Command {
+func (s *CometBFTServer[AppT, T]) ShowValidatorCmd() *cobra.Command {
 	cmd := cobra.Command{
 		Use:   "show-validator",
 		Short: "Show this node's CometBFT validator info",
@@ -135,7 +134,7 @@ func (s *CometBFTServer[T]) ShowValidatorCmd() *cobra.Command {
 }
 
 // ShowAddressCmd - show this node's validator address
-func (s *CometBFTServer[T]) ShowAddressCmd() *cobra.Command {
+func (s *CometBFTServer[AppT, T]) ShowAddressCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "show-address",
 		Short: "Shows this node's CometBFT validator consensus address",
@@ -154,7 +153,7 @@ func (s *CometBFTServer[T]) ShowAddressCmd() *cobra.Command {
 }
 
 // VersionCmd prints CometBFT and ABCI version numbers.
-func (s *CometBFTServer[T]) VersionCmd() *cobra.Command {
+func (s *CometBFTServer[AppT, T]) VersionCmd() *cobra.Command {
 	return &cobra.Command{
 		Use:   "version",
 		Short: "Print CometBFT libraries' version",
@@ -182,7 +181,7 @@ func (s *CometBFTServer[T]) VersionCmd() *cobra.Command {
 }
 
 // QueryBlocksCmd returns a command to search through blocks by events.
-func (s *CometBFTServer[T]) QueryBlocksCmd() *cobra.Command {
+func (s *CometBFTServer[AppT, T]) QueryBlocksCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "blocks",
 		Short: "Query for paginated blocks that match a set of events",
@@ -203,8 +202,8 @@ for. Each module documents its respective events under 'xx_events.md'.
 			}
 
 			query, _ := cmd.Flags().GetString(auth.FlagQuery)
-			page, _ := cmd.Flags().GetInt(flags.FlagPage)
-			limit, _ := cmd.Flags().GetInt(flags.FlagLimit)
+			page, _ := cmd.Flags().GetInt(FlagPage)
+			limit, _ := cmd.Flags().GetInt(FlagLimit)
 			orderBy, _ := cmd.Flags().GetString(auth.FlagOrderBy)
 
 			blocks, err := rpc.QueryBlocks(cmd.Context(), rpcclient, page, limit, query, orderBy)
@@ -221,9 +220,9 @@ for. Each module documents its respective events under 'xx_events.md'.
 		},
 	}
 
-	flags.AddQueryFlagsToCmd(cmd)
-	cmd.Flags().Int(flags.FlagPage, query.DefaultPage, "Query a specific page of paginated results")
-	cmd.Flags().Int(flags.FlagLimit, query.DefaultLimit, "Query number of transactions results per page returned")
+	AddQueryFlagsToCmd(cmd)
+	cmd.Flags().Int(FlagPage, query.DefaultPage, "Query a specific page of paginated results")
+	cmd.Flags().Int(FlagLimit, query.DefaultLimit, "Query number of transactions results per page returned")
 	cmd.Flags().String(auth.FlagQuery, "", "The blocks events query per CometBFT's query semantics")
 	cmd.Flags().String(auth.FlagOrderBy, "", "The ordering semantics (asc|dsc)")
 	_ = cmd.MarkFlagRequired(auth.FlagQuery)
@@ -232,7 +231,7 @@ for. Each module documents its respective events under 'xx_events.md'.
 }
 
 // QueryBlockCmd implements the default command for a Block query.
-func (s *CometBFTServer[T]) QueryBlockCmd() *cobra.Command {
+func (s *CometBFTServer[AppT, T]) QueryBlockCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "block --type=[height|hash] [height|hash]",
 		Short: "Query for a committed block by height, hash, or event(s)",
@@ -312,14 +311,14 @@ $ %s query block --%s=%s <hash>
 		},
 	}
 
-	flags.AddQueryFlagsToCmd(cmd)
+	AddQueryFlagsToCmd(cmd)
 	cmd.Flags().String(auth.FlagType, auth.TypeHash, fmt.Sprintf("The type to be used when querying tx, can be one of \"%s\", \"%s\"", auth.TypeHeight, auth.TypeHash))
 
 	return cmd
 }
 
 // QueryBlockResultsCmd implements the default command for a BlockResults query.
-func (s *CometBFTServer[T]) QueryBlockResultsCmd() *cobra.Command {
+func (s *CometBFTServer[AppT, T]) QueryBlockResultsCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "block-results [height]",
 		Short: "Query for a committed block's results by height",
@@ -364,7 +363,7 @@ func (s *CometBFTServer[T]) QueryBlockResultsCmd() *cobra.Command {
 		},
 	}
 
-	flags.AddQueryFlagsToCmd(cmd)
+	AddQueryFlagsToCmd(cmd)
 
 	return cmd
 }
@@ -384,7 +383,7 @@ func parseOptionalHeight(heightStr string) (*int64, error) {
 	return &tmp, nil
 }
 
-func (s *CometBFTServer[T]) BootstrapStateCmd() *cobra.Command {
+func (s *CometBFTServer[AppT, T]) BootstrapStateCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "bootstrap-state",
 		Short: "Bootstrap CometBFT state at an arbitrary block height using a light client",
@@ -396,7 +395,7 @@ func (s *CometBFTServer[T]) BootstrapStateCmd() *cobra.Command {
 				return err
 			}
 			if height == 0 {
-				height, err = s.App.store.GetLatestVersion()
+				height, err = s.Consensus.store.GetLatestVersion()
 				if err != nil {
 					return err
 				}
@@ -414,7 +413,7 @@ func (s *CometBFTServer[T]) BootstrapStateCmd() *cobra.Command {
 
 func printOutput(cmd *cobra.Command, out []byte) error {
 	// Get flags output
-	outFlag, err := cmd.Flags().GetString(flags.FlagOutput) 
+	outFlag, err := cmd.Flags().GetString(FlagOutput)
 	if err != nil {
 		return err
 	}